Add structure for encoding/decoding parameters, errors, events, return values using schema
